requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-anthropic
anthropic
python-dotenv
requests